# internal/httpapi/openapi.yaml
openapi: 3.0.3
info:
  title: Countinghouse API
  description: >
    Read-side energy cost/accounting HTTP API. Turns per-device telemetry into
    windowed kWh and cost. All endpoints except /healthz and /openapi.json
    require a Bearer JWT (user or service token) issued by the configured
    identity service.
  version: "0.2.0"
  contact:
    name: Martin Sweeney
    email: martin@sweeney.io
    url: https://github.com/sweeney/countinghouse
  license:
    name: Proprietary
    url: https://github.com/sweeney/countinghouse
servers:
  - url: __PUBLIC_URL__
security:
  - bearerAuth: []
...
paths:
...
  /devices/{id}/energy:
    get:
      summary: Device energy
      description: >
        Windowed kWh consumed by one device. Routing is class-derived: plug
        classes use the reset-safe counter path (source "counter"); ups_sensor
        uses the power integral (source "integral"). Non-metered classes return
        400.
      operationId: getDeviceEnergy
      tags: [data]
      parameters:
        - $ref: "#/components/parameters/DeviceID"
        - $ref: "#/components/parameters/Window"
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
      responses:
        "200":
          description: Windowed energy for the device
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceEnergyResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /devices/{id}/cost:
    get:
      summary: Device cost
      description: >
        Windowed kWh and VAT-inclusive GBP cost for one device, at the tariff
        effective at the window start.
      operationId: getDeviceCost
      tags: [data]
      parameters:
        - $ref: "#/components/parameters/DeviceID"
        - $ref: "#/components/parameters/Window"
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
      responses:
        "200":
          description: Windowed cost for the device
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceCostResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/Unavailable"
  /series:
    get:
      summary: Energy time-series
      description: >
        Multi-series energy time-series over the window: per-series kWh, cost and
        avg_w. group_by selects device (default, one series per metered device,
        excluding the whole-house meter), room, class, or house. house yields
        three series in order: "monitored" = sum of devices, "unmonitored" =
        clamp(meter − monitored) per bucket (the rest-of-home energy no monitored
        device accounts for), and "meter" = the whole-house meter. Stacking every
        series except "meter" reconstructs the whole home; monitored + unmonitored
        == meter per bucket within rounding tolerance (a slightly-negative residual
        from counter quantisation clamps to 0). unmonitored cost is the tariff
        applied to its clamped energy. The house response also carries top-level
        coverage (monitored ÷ meter) and a stale-monitored-device indicator
        (count + ids) so consumers can gauge how trustworthy the split is. For
        device/room/class groupings, include_unmonitored=true adds the same
        rest-of-home quantity as a single catch-all series so those parts sum to
        the meter too. The shape param chooses columnar (default) or row-oriented
        output.
      operationId: getSeries
      tags: [data]
      parameters:
        - $ref: "#/components/parameters/Window"
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
        - $ref: "#/components/parameters/Interval"
        - $ref: "#/components/parameters/GroupBy"
        - $ref: "#/components/parameters/IncludeUnmonitored"
        - $ref: "#/components/parameters/Unclamped"
        - $ref: "#/components/parameters/Shape"
      responses:
        "200":
          description: >
            Energy time-series. shape=columns → SeriesResponse (shared buckets
            axis + per-series arrays); shape=rows → RowsResponse (flat rows).
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/SeriesResponse"
                  - $ref: "#/components/schemas/RowsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "502":
          $ref: "#/components/responses/BadGateway"
        "503":
          $ref: "#/components/responses/Unavailable"
  /devices/{id}/series:
    get:
      summary: Device energy time-series
      description: >
        Single-device convenience form of /series: the same response shapes
        carrying exactly one series for the device. Non-metered classes return
        400. Every energy-capable device in /devices is served, INCLUDING the
        whole-house meter: it is excluded from /series?group_by=device because
        it would double-count against the plugs it measures, but a request for a
        single device cannot double-count, and its series here is identical to
        the "meter" series of group_by=house. The response always carries exactly
        one series and group_by reads "device"; a build that produced none is an
        internal inconsistency and returns 500 rather than 200 with an empty
        series list. The shape param chooses columnar (default) or row-oriented
        output.
        The reserved id "unmonitored" is a synthetic device serving the house
        "unmonitored" series (rest of home = clamp(meter − monitored)) through this
        same single-series shape; it returns 404 when no whole-house meter is
        configured (unmonitored is then undefined). For id=unmonitored the
        unclamped diagnostic param applies (ignored for real devices).
      operationId: getDeviceSeries
      tags: [data]
      parameters:
        - $ref: "#/components/parameters/DeviceID"
        - $ref: "#/components/parameters/Window"
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
        - $ref: "#/components/parameters/Interval"
        - $ref: "#/components/parameters/Unclamped"
        - $ref: "#/components/parameters/Shape"
      responses:
        "200":
          description: >
            Energy time-series for the device. shape=columns → SeriesResponse;
            shape=rows → RowsResponse.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/SeriesResponse"
                  - $ref: "#/components/schemas/RowsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Internal"
        "502":
          $ref: "#/components/responses/BadGateway"
        "503":
          $ref: "#/components/responses/Unavailable"
  /bill:
    get:
      summary: Electricity bill
      description: >
        Per-device cost breakdown plus standing charge, total, and
        reconciliation against the whole-house meter for the window.
      operationId: getBill
      tags: [data]
      parameters:
        - $ref: "#/components/parameters/Window"
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
      responses:
        "200":
          description: Assembled bill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/Unavailable"
  /tariffs:
    get:
      summary: Current tariffs by fuel
      description: >
        All configured tariffs keyed by fuel (electricity, gas, ...). Rates are
        ex-VAT; vat_rate gives the gross-up. Countinghouse bills electricity
        today; gas is exposed for forward-compatibility with gas devices.
      operationId: getTariffs
      tags: [data]
      responses:
        "200":
          description: Tariffs keyed by fuel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TariffsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /devices:
    get:
      summary: Device catalog
      description: >
        Discovery catalog: a pass-through of the statehouse_devices namespace
        enriched with a derived capabilities hint ("energy" when the class is
        metered, "events" for event-bearing classes — binary_state_device,
        fire_alarm). Lets a UI build a device picker without knowing Influx.
        Sorted by id. When a whole-house meter is configured, a synthetic
        "unmonitored" entry (class "unmonitored", capability "energy", id reserved)
        is included so the rest-of-home series appears in pickers and plots through
        /devices/unmonitored/series like any device.
      operationId: getDevices
      tags: [data]
      responses:
        "200":
          description: Device catalog
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DevicesCatalog"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /floors:
    get:
      summary: Floor catalog
      description: >
        The building's floors, and the discoverable vocabulary behind the
        `floors=` filter and `group_by=floor`. Lists exactly the floors at least
        one BILLED device declares (metered, excluding the whole-house meter),
        which is exactly the set `floors=` accepts — so a picker filled from this
        endpoint cannot build a request /series rejects. A floorplan record for a
        floor with no metered device is therefore not listed, and a floor devices
        declare that the floorplan has no record for IS listed, with its name
        empty and order null. Sorted by declared storey order ascending, then by
        id; floors with no declared order sort last. Reads the same floorplan
        namespace greenhouse serves /floors from, so both services answer in one
        vocabulary.
      operationId: getFloors
      tags: [data]
      responses:
        "200":
          description: Floor catalog
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FloorsCatalog"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /rooms:
    get:
      summary: Room catalog
      description: >
        The building's rooms, and the discoverable vocabulary behind the `rooms=`
        filter and `group_by=room`. Listed on the same rule as /floors: exactly
        the rooms at least one billed device is attributed to. `category` is
        relayed RAW rather than reduced to a flag — a plant room's consumption is
        infrastructure rather than household usage, but which rooms "count" is a
        per-dashboard policy question. The reserved "house" key (a coverage
        scope, not a room) never appears. Sorted by floor storey order, then by
        room id; rooms whose floor is unknown sort last.
      operationId: getRooms
      tags: [data]
      responses:
        "200":
          description: Room catalog
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomsCatalog"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /devices/{id}/events:
    get:
      summary: Device event timeline
      description: >
        Ordered state-transition edges for one device over the window
        (vertical-line overlays). Each event carries the raw from/to labels and
        a normalized on flag (omitted for momentary/unknown labels).
      operationId: getDeviceEvents
      tags: [data]
      parameters:
        - $ref: "#/components/parameters/DeviceID"
        - $ref: "#/components/parameters/Window"
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
      responses:
        "200":
          description: Transition events for the device
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceEventsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          $ref: "#/components/responses/BadGateway"
  /devices/{id}/intervals:
    get:
      summary: Device on/off intervals
      description: >
        Sustained ON spans for one device over the window (shaded bands), with
        the carry-in opening state (state_at_start), per-interval durations, an
        open flag for an in-progress span at the window end, and summary stats
        (on_count, total_on_seconds, duty).
      operationId: getDeviceIntervals
      tags: [data]
      parameters:
        - $ref: "#/components/parameters/DeviceID"
        - $ref: "#/components/parameters/Window"
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
      responses:
        "200":
          description: Intervals and stats for the device
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceIntervalsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          $ref: "#/components/responses/BadGateway"
  /events:
    get:
      summary: Multi-device event overlay
      description: >
        Categorical timeline overlay across a device set. The set comes from
        devices= (csv of ids; an unknown id is a 400), class= (all devices of a
        class), or the default (all event-bearing devices: binary_state_device,
        fire_alarm). group_by=device
        (default) yields one series per device; group_by=class merges every
        member device's events under each class key, time-ordered.
      operationId: getEvents
      tags: [data]
      parameters:
        - name: devices
          in: query
          required: false
          description: Comma-separated device ids to overlay. An unknown id is a 400.
          schema:
            type: string
        - name: class
          in: query
          required: false
          description: Restrict the set to devices of this class (ignored when devices= is given).
          schema:
            type: string
        - $ref: "#/components/parameters/Window"
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
        - name: group_by
          in: query
          required: false
          description: How to group series. Defaults to "device".
          schema:
            type: string
            enum: [device, class]
            default: device
      responses:
        "200":
          description: Columnar multi-device event overlay
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "502":
          $ref: "#/components/responses/BadGateway"
components:
  parameters:
    DeviceID:
      name: id
      in: path
      required: true
      description: Device id (key in the statehouse_devices namespace).
      schema:
        type: string
    Window:
      name: window
      in: query
      required: false
      description: >-
        Window spec. Defaults to "today". Period-to-date: today, week (from
        Monday), month (from the 1st) — all ending now. custom: an explicit
        from/to range. Rolling: "<N>d" is a trailing N calendar days ending now,
        DAY-ALIGNED to local midnight (e.g. 7d = today + the previous 6 days;
        1d == today); "<N>h" is an EXACT trailing N hours (e.g. 24h), not
        midnight-aligned. N is a positive integer.
      schema:
        type: string
        default: today
        pattern: '^(today|week|month|custom|[0-9]+[dh])$'
        example: 7d
    From:
      name: from
      in: query
      required: false
      description: >-
        Custom window start (RFC3339). Required when window=custom, and ONLY
        valid then — supplying from with any other window (period-to-date
        today/week/month or rolling <N>d/<N>h, which derive their own range) is
        rejected with 400.
      schema:
        type: string
        format: date-time
    To:
      name: to
      in: query
      required: false
      description: >-
        Custom window end (RFC3339). Required when window=custom, and ONLY
        valid then — supplying to with any other window (period-to-date
        today/week/month or rolling <N>d/<N>h, which derive their own range) is
        rejected with 400.
      schema:
        type: string
        format: date-time
    Interval:
      name: interval
      in: query
      required: false
      description: >
        Bucket granularity. Defaults smartly to the window (today→1h,
        week/month→1d, custom and rolling <N>d/<N>h→by span). An interval whose bucket count over the
        window exceeds the ~1000 cap returns 400 asking for a coarser interval.
        For window=custom whose from is not on an interval boundary, the bucket
        axis snaps DOWN to the interval grid anchored at local midnight (so it
        matches the aggregation boundaries); the first bucket is widened to that
        boundary. Period-to-date windows already start at local midnight and are
        unaffected.
      schema:
        type: string
        enum: [5m, 15m, 30m, 1h, 6h, 1d]
    GroupBy:
      name: group_by
      in: query
      required: false
      description: How to group series. Defaults to "device".
      schema:
        type: string
        enum: [device, room, class, house]
        default: device
    IncludeUnmonitored:
      name: include_unmonitored
      in: query
      required: false
      description: >
        When true, a device/room/class grouping gains exactly one extra
        "unmonitored" (rest-of-home) series so the parts sum to the whole-house
        meter. Never subdivided (one series regardless of grouping); a no-op when
        no meter is configured. Ignored for group_by=house (which always includes
        it). Default false — the response is otherwise unchanged.
      schema:
        type: boolean
        default: false
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Missing or invalid Bearer JWT
      headers:
        WWW-Authenticate:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Invalid or missing query parameters, or device has no energy series
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Unknown device or absent tariff
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unavailable:
      description: No tariff configured
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    DeviceEnergyResponse:
      type: object
      properties:
        device_id:
          type: string
        kwh:
          type: number
        source:
          type: string
          enum: [counter, integral]
        window:
          type: string
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
      required: [device_id, kwh, source, window, from, to]
    DeviceCostResponse:
      type: object
      properties:
        device_id:
          type: string
        kwh:
          type: number
        cost:
          type: number
        currency:
          type: string
          example: GBP
        window:
          type: string
        tariff:
          type: object
          properties:
            unit_rate:
              type: number
            vat_rate:
              type: number
      required: [device_id, kwh, cost, currency, window]
...